requirements.txt
python-dotenv
openai
azure-ai-openai
azure-search-documents
langchain
ragas
pandas
requests
datasets